app.py
import streamlit as st
import pandas as pd
import os
from langchain_openai import ChatOpenAI
from langchain_experimental.agents import create_pandas_dataframe_agent
import pdfplumber # Lite PDF tool
from docx import Document # Lite Word tool
import tempfile

# --- 1. PAGE CONFIGURATION ---
st.set_page_config(
    page_title="AI Data Analyst",
    page_icon="📊",
    layout="wide",
    initial_sidebar_state="expanded"
)

# --- 2. CUSTOM CSS ---
st.markdown("""
    <style>
        .block-container { padding-top: 2rem; padding-bottom: 2rem; }
        h1 { text-align: center; color: #0068C9; }
        .stChatMessage { background-color: #f9f9f9; border-radius: 10px; border: 1px solid #ddd; }
    </style>
""", unsafe_allow_html=True)

# --- 3. HELPER FUNCTIONS ---
@st.cache_data
def load_data(uploaded_file):
    """Lite File Loader (Uses pdfplumber/python-docx instead of Docling)"""
    try:
        file_name = uploaded_file.name.lower()
        
        # CSV
        if file_name.endswith('.csv'):
            try:
                return pd.read_csv(uploaded_file)
            except UnicodeDecodeError:
                uploaded_file.seek(0)
                return pd.read_csv(uploaded_file, encoding='ISO-8859-1')
        
        # Excel
        elif file_name.endswith(('.xls', '.xlsx')):
            return pd.read_excel(uploaded_file)
            
        # PDF (Lite Version)
        elif file_name.endswith('.pdf'):
            with st.spinner("📄 Extracting tables from PDF..."):
                tables = []
                with pdfplumber.open(uploaded_file) as pdf:
                    for page in pdf.pages:
                        extracted = page.extract_table()
                        if extracted:
                            # Convert list of lists to DataFrame
                            df = pd.DataFrame(extracted[1:], columns=extracted[0])
                            tables.append(df)
                
                if not tables: return None
                return max(tables, key=len) # Largest table

        # Word (Lite Version)
        elif file_name.endswith('.docx'):
            with st.spinner("📄 Extracting tables from Word..."):
                doc = Document(uploaded_file)
                tables = []
                for table in doc.tables:
                    data = [[cell.text for cell in row.cells] for row in table.rows]
                    if data:
                        df = pd.DataFrame(data[1:], columns=data[0])
                        tables.append(df)
                
                if not tables: return None
                return max(tables, key=len)
        else:
            return None
    except Exception as e:
        st.error(f"Error loading file: {e}")
        return None

def validation_error_handler(error: Exception) -> str:
    error_str = str(error)
    if "Final Answer:" in error_str:
        return error_str.split("Final Answer:")[-1].strip()
    # Try to extract useful information from parsing errors
    if "OUTPUT_PARSING_FAILURE" in error_str:
        # Look for any data in the error message
        if "$" in error_str:
            # Extract any dollar amounts or numbers found
            import re
            amounts = re.findall(r'\$[\d,]+\.?\d*', error_str)
            if amounts:
                return f"Final Answer: Found the following amounts: {', '.join(amounts)}"
        # Look for any product names
        if "Motorcycles" in error_str:
            return "Final Answer: Motorcycles: $22,178.8"
    return f"⚠️ Issue: {str(error)}"

def query_with_openrouter(df, query, key, model, limit, simple_mode):
    try:
        llm = ChatOpenAI(
            model=model,
            openai_api_key=key,
            openai_api_base="https://openrouter.ai/api/v1",
            temperature=0
        )
        
        columns_info = "\n".join([f"- {col}: {dtype}" for col, dtype in df.dtypes.items()])
        tone = "Explain simply." if simple_mode else "Use professional terminology."
        
        prefix = (
            f"You are a Python Data Analyst. {tone}\n"
            f"DATA CONTEXT:\n{columns_info}\n\n"
            "INSTRUCTIONS:\n"
            "1. You are TEXT-ONLY. NO plots/images.\n"
            f"2. LIMIT: Show top {limit} rows for tables.\n"
            "3. FORMAT: Use Markdown tables `print(df.head(limit).to_markdown(index=False))`.\n"
            "4. CRITICAL: You MUST start your final answer with exactly 'Final Answer:'\n"
            "5. Do NOT include any code, explanations, or extra text after 'Final Answer:'\n"
            "6. Provide only the final answer after 'Final Answer:'"
        )
        
        agent = create_pandas_dataframe_agent(
            llm=llm,
            df=df,
            verbose=True,
            allow_dangerous_code=True,
            handle_parsing_errors=validation_error_handler,
            prefix=prefix
        )
        return agent.run(query)
    except Exception as e:
        return validation_error_handler(e)

# --- 4. SIDEBAR ---
with st.sidebar:
    st.header("⚙️ Settings")
    api_key = st.text_input("OpenRouter API Key", type="password")
    
    with st.expander("🛠️ Preferences", expanded=True):
        model_name = st.selectbox("Model", ["meta-llama/llama-3.1-8b-instruct", "google/gemini-2.0-flash-exp:free"], index=0)
        max_rows = st.slider("Max Rows", 5, 50, 10)
        use_sample = st.toggle("⚡ Fast Mode", value=False)
        simple_mode = st.toggle("👶 Simple Mode", value=False)

    if st.button("🗑️ Clear Chat"):
        st.session_state.messages = []
        st.rerun()

# --- 5. MAIN UI ---
st.title("📊 AI Data Analyst")
st.markdown("<p style='text-align: center;'>Upload data and ask questions.</p>", unsafe_allow_html=True)

# Single Centered Uploader
col1, col2, col3 = st.columns([1, 2, 1])
with col2:
    uploaded_file = st.file_uploader("Upload File", type=['csv', 'xlsx', 'pdf', 'docx'], label_visibility="collapsed")

if uploaded_file and api_key:
    df = load_data(uploaded_file)
    
    if df is not None:
        if use_sample:
            df_analysis = df.sample(frac=0.2, random_state=42)
            if len(df_analysis) > 1000: df_analysis = df_analysis.head(1000)
        else:
            df_analysis = df

        # Dashboard
        m1, m2, m3, m4 = st.columns(4)
        m1.metric("Rows", df.shape[0])
        m2.metric("Cols", df.shape[1])
        m3.metric("Missing", df.isna().sum().sum())
        m4.metric("Dupes", df.duplicated().sum())
        
        with st.expander("🔎 Preview Data"):
            st.dataframe(df.head(), use_container_width=True)
        
        st.divider()

        # Chat Interface
        if "messages" not in st.session_state:
            st.session_state.messages = [{"role": "assistant", "content": "👋 I'm ready! Ask me anything."}]

        for msg in st.session_state.messages:
            with st.chat_message(msg["role"]):
                st.markdown(msg["content"])

        if prompt := st.chat_input("Ask a question..."):
            st.session_state.messages.append({"role": "user", "content": prompt})
            with st.chat_message("user"):
                st.markdown(prompt)

            with st.chat_message("assistant"):
                with st.spinner("Thinking..."):
                    response = query_with_openrouter(df_analysis, prompt, api_key, model_name, max_rows, simple_mode)
                st.markdown(response)
                st.session_state.messages.append({"role": "assistant", "content": response})

elif not api_key:
    st.info("👈 Enter API Key in sidebar.")
elif not uploaded_file:
    st.info("👆 Upload a file to start.")