requirements.txt
streamlit
pandas
langchain-openai
langchain-experimental
openpyxl
tabulate
pdfplumber
python-docx
